+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "hivelocal-hk",
-    "version": "0.1.5",
+    "version": "0.1.6",
     "private": true,
     "scripts": {
         "dev": "next dev",
