--- src/components/map/Box.tsx
@@ -1,42 +1,0 @@
-// import React from "react";
-
-// /* Don't forget to download the CSS file too
-// OR remove the following line if you're already using Tailwind */
-
-// import "./style.css";
-
-// export const MyPlugin = () => {
-//     return (
-//         <div id='webcrumbs'>
-//             <div className='w-[400px] bg-white rounded-xl shadow-lg p-8 hover:shadow-xl transition-all duration-300'>
-//                 <header className='mb-6'>
-//                     <div className='bg-gray-50 p-4 rounded-lg hover:bg-gray-100 transition-all duration-300'>
-//                         <h2 className='text-3xl font-bold mb-2'>Central Park Concert</h2>
-//                         <div className='border-t border-gray-200 mt-2 pt-2'>
-//                             <p className='text-lg'>NYC Parks Department</p>
-//                         </div>
-//                     </div>
-//                 </header>
-
-//                 <div className='space-y-6'>
-//                     <div className='flex justify-between items-center bg-gray-50 p-4 rounded-lg hover:bg-gray-100 transition-all duration-300'>
-//                         <div className='w-full text-center'>
-//                             <p className='text-lg font-medium'>15 Aug 2023 - 17 Aug 2023</p>
-//                             <p className='text-lg mt-2'>6:00 PM - 10:00 PM</p>
-//                         </div>
-//                     </div>
-
-//                     <div className='bg-gray-50 p-4 rounded-lg hover:bg-gray-100 transition-all duration-300'>
-//                         <p className='text-lg font-medium mb-1'>District</p>
-//                         <p className='text-lg'>Manhattan</p>
-//                     </div>
-
-//                     <div className='bg-gray-50 p-4 rounded-lg hover:bg-gray-100 transition-all duration-300'>
-//                         <p className='text-lg font-medium mb-1'>Address</p>
-//                         <p className='text-lg'>Central Park Mall, New York, NY 10022</p>
-//                     </div>
-//                 </div>
-//             </div>
-//         </div>
-//     );
-// };
