+++ src/components/sidebar/LocationCard.tsx
